Error handling
Reorganize the whole package into three files: `main.rs`, `game.rs` and `error.rs`.

Create a specific error type for this application, that contains a "terminal size" error type, and correctly breaks if the terminal isn't fit for the proposed game.

I'm using a macro to remove some code repetition on the error handling, but I'm not sure if I liked the actual solution.